Add files via upload
process_excel.py is used to read data in Google Sheet and send it to a language model e.g Claude. The answer is then written back to a new tab in the same Google Sheet.
cal_bert_rouge.py is used to evaluate the answer given by the language model using Rouge L and BertScore. The scores are written back to the Sheet.